Two-column layout, types.ts -> api-types.ts
